microservices/job_offers_scrapper/praca/requirements.txt
requests
google-cloud-storage
beautifulsoup4
uvicorn
fastapi
pydantic